added app page layout
