+++ live_monitoring_screen/App.js
@@ -1,21 +1,23 @@
 import { StatusBar } from 'expo-status-bar';
-import React from 'react';
+import React from "react";
+import "bootstrap/dist/css/bootstrap.min.css";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { StyleSheet, Text, View } from 'react-native';
 
-export default function App() {
+import BedNumberPage from "./pages/bedNumberPage";
+import FallRiskStatusPage from "./pages/fallRiskStatusPage"
+import ReadyPage from "./pages/readyPage"
+
+const AppRouter = () => {
   return (
-    <View style={styles.container}>
-      <Text>Open up App.js to start working on your app!</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+    <Router>
+      <Switch>
+        <Route exact path = "/" component = {ReadyPage} />
+        <Route exact path = "/bednumber" component = {BedNumberPage} />
+        <Route exact path = "/fallriskstatus" component = {FallRiskStatusPage} />
+      </Switch>
+    </Router>
+  )
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+export default AppRouter;
