+++ live_monitoring_screen/index.css
@@ -1,0 +1,40 @@
+body {
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+      sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+  }
+  
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+      monospace;
+  }
+  
+  /* Initialise fonts */
+  @font-face {
+    font-family: 'Gotham-Thin';
+    src: url(./Fonts/Gotham-Thin.otf) format('opentype');
+  }
+  
+  @font-face {
+    font-family: 'Gotham-Medium';
+    src: url(./Fonts/Gotham-Medium.otf) format('opentype');
+  }
+  
+  @font-face {
+    font-family: 'Gotham-Black';
+    src: url(./Fonts/Gotham-Black.otf) format('opentype');
+  }
+  
+  @font-face {
+    font-family: 'Gotham-Ultra';
+    src: url(./Fonts/Gotham-Ultra.otf) format('opentype');
+  }
+  
+  @font-face {
+    font-family: 'Gotham-Book';
+    src: url(./Fonts/Gotham-Book.otf) format('opentype');
+  }
+
